Separate the concept of Transform and GameObject; separate the concept of SceneManager and Scene
