+++ src/game_engine/unity/scene_manager/offsets.rs
@@ -9,6 +9,8 @@ pub(super) struct Offsets {
     pub(super) root_storage_container: u8,
     pub(super) game_object: u8,
     pub(super) game_object_name: u8,
+    pub(super) game_object_activeself: u8,
+    pub(super) game_object_activeinhierarchy: u8,
     pub(super) klass: u8,
     pub(super) klass_name: u8,
     pub(super) children_pointer: u8,
@@ -26,6 +28,8 @@ impl Offsets {
                 root_storage_container: 0xB0,
                 game_object: 0x30,
                 game_object_name: 0x60,
+                game_object_activeself: 0x5E,
+                game_object_activeinhierarchy: 0x5F,
                 klass: 0x28,
                 klass_name: 0x48,
                 children_pointer: 0x70,
@@ -39,6 +43,8 @@ impl Offsets {
                 root_storage_container: 0x88,
                 game_object: 0x1C,
                 game_object_name: 0x3C,
+                game_object_activeself: 0x32,
+                game_object_activeinhierarchy: 0x33,
                 klass: 0x18,
                 klass_name: 0x2C,
                 children_pointer: 0x50,
